Part 4.1. - 4.23.

+++ Osa_4/blogilistasovellus/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blogs')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+      .find({}).populate('user', {username: 1, name: 1})
     response.json(blogs)
   })
 
@@ -18,18 +21,26 @@ blogRouter.get('/', async (request, response) => {
     response.json(updateBlog)
   })
   
-blogRouter.post('/', (request, response, next) => {
+blogRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  if (!request.body.title || !request.body.url) {
+    return response.status(400).end()
+  }
+
    const blog = new Blog({
     ...request.body,
-    likes: request.body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
   
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
-    .catch(error => next(error))
+  const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogRouter.get('/:id', async (request, response) => {
@@ -41,9 +52,22 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', userExtractor,async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user
+
+  
+  if (!blog) {
+    return response.status(401).json({ error: 'no blog found'})
+  }
+
+  if (blog.user.toString() !== user.id) {
+    return response.status(403).json({ error: 'not authorized to delete this blog' });
+  } 
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
+
 })
